src/requirements.txt
requests
python-dotenv
fastapi
uvicorn[standard]
apscheduler
python-jose[cryptography]
razorpay